+++ changelog.md
@@ -1,4 +1,7 @@
 
+**<span style="color:#56adda">0.0.3</span>**
+- Add smart output target to Basic mode so matching audio codecs can be re-encoded when they are significantly above the recommended bitrate window
+
 **<span style="color:#56adda">0.0.2</span>**
 - Update runner signatures to accept keyword helper args for Unmanic compatibility